BUG: Fix crash in registering markups displ. man. when connecting to device
When registering the markups displayable manager, and then it creates widgets for the existing markups, it wants to use the application logic to get the markups logic, which it is supposed to get from the factory, but since in the factory it was not set, getting the markups logic failed in absence of the application logic and it caused a crash. Fixed by setting the application logic directly.

This commit is based on similar changes contributed
in KitwareMedical/SlicerVirtualReality@5b6b1b9f0 (ENH: Convert the interactor style to use the new vtkMRMLViewInteractorStyle)